+++ extensions/com_pungamail/administrator/components/com_pungamail/tmpl/subscribers/default.php
@@ -22,7 +22,6 @@ $statusLabels = [
 <?php echo \Joomla\CMS\Layout\LayoutHelper::render('pungamail.section_navigation', ['section' => 'audience', 'active' => 'subscribers'], JPATH_ADMINISTRATOR . '/components/com_pungamail/layouts'); ?>
 <form action="<?php echo Route::_(AdministratorRoute::subscribers()); ?>" method="post" name="adminForm" id="adminForm">
 	<?php echo LayoutHelper::render('joomla.searchtools.default', ['view' => $this]); ?>
-	<p class="small text-muted mb-3"><?php echo Text::_('COM_PUNGAMAIL_SUPPRESSION_HELP'); ?></p>
 
 	<div class="table-responsive">
 		<table class="table itemList" id="subscriberList">